syntax: Generate type groups for "parsing" of type lists vs single types
Should help when highlighting function arguments with type annotations
since we can now distinguish between a type annotation of a single type
or a type annotation of a list of types.

This also allows for a refactor of how the function signature group
worked and got rid of some annoying to work around code.